+++ lib/src/channels/channels.dart
@@ -4,8 +4,11 @@
 library youtube_explode.channels;
 
 export 'channel.dart';
+export 'channel_about.dart';
 export 'channel_client.dart';
 export 'channel_id.dart';
+export 'channel_link.dart';
+export 'channel_thumbnail.dart';
 export 'channel_video.dart';
 export 'username.dart';
 export 'video_sorting.dart';