+++ lib/src/exceptions/video_unplayable_exception.dart
@@ -28,7 +28,4 @@ class VideoUnplayableException implements YoutubeExplodeException {
   VideoUnplayableException.notLiveStream(VideoId videoId)
       : message = 'Video \'$videoId\' is not an ongoing live stream.\n'
             'Live stream manifest is not available for this video';
-
-  @override
-  String toString() => 'VideoUnplayableException: $message';
 }