+++ CHANGELOG.md
@@ -3,6 +3,14 @@
 - Only throw custom exceptions from the library.
 - `getUploadsFromPage` no longer throws.
 
+## 1.6.0
+- BREAKING CHANGE: Renamed `getVideosAsync` to `getVideos`.
+- Implemented `getVideosFromPage` which supersedes `queryFromPage`.
+- Implemented JSON Classes for reverse engineer.
+- Added `forceWatchPage` to the video client to assure the fetching of the video page. (ATM useful only if using the comments api)
+- Remove adaptive streams. These are not used anymore.
+- Implement `channelClient.getAboutPage` and `getAboutPageByUsername` to fetch data from a channel's about page.
+
 ## 1.5.2
 - Fix extraction for same videos (#76)
 
@@ -11,7 +19,6 @@
 
 ## 1.5.0
 - BREAKING CHANGE: Renamed `Container` class to `StreamContainer` to avoid conflicting with Flutter `Container`. See #66
-
 
 ## 1.4.4
 - Expose HttpClient in APIs